+++ ODPC.Server/Features/Gebruikersgroep/GebruikersgroepDetails/GebruikersgroepDetailsController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ODPC.Data;
+using ODPC.Data.Entities;
 
 namespace ODPC.Features.Gebruikersgroep.GebruikersgroepDetails
 {
@@ -13,23 +14,9 @@ namespace ODPC.Features.Gebruikersgroep.GebruikersgroepDetails
         [HttpGet("api/v1/gebruikersgroepen/{uuid:guid}")]
         public async Task<IActionResult> Get(Guid uuid, CancellationToken token)
         {
-            var groep = await _context.Gebruikersgroepen.SingleOrDefaultAsync(x => x.Uuid == uuid, cancellationToken: token);
+            var groep = await _context.Gebruikersgroepen.Include(x=>x.Waardelijsten).SingleOrDefaultAsync(x => x.Uuid == uuid, cancellationToken: token);
 
-            if (groep == null) return NotFound();
-
-            var result = new GebruikersgroepDetailsModel
-            {
-                Uuid = groep.Uuid,
-                Naam = groep.Naam,
-                Omschrijving = groep.Omschrijving,
-                GekoppeldeWaardelijsten = await _context
-                    .GebruikersgroepWaardelijsten
-                    .Where(x => x.GebruikersgroepUuid == uuid)
-                    .Select(x => x.WaardelijstId)
-                    .ToListAsync(cancellationToken: token)
-            };
-
-            return Ok(result);
+            return groep == null ? NotFound() : Ok(GebruikersgroepDetailsModel.MapEntityToViewModel(groep));
         }
     }
 }
